add user tags change level and tests, activity add tags (version 0.001)

+++ src/main/java/com/angrydwarfs/framework/repository/LevelRepository.java
@@ -24,6 +24,6 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface LevelRepository extends JpaRepository<Level, Long> {
+public interface LevelRepository extends JpaRepository<Level, Integer> {
     Optional<Level> findByLevelName(ELevel levelName);
 }
